Namespace components
So they are easier to find when we have components for Collections too
