--- app/components/deposit_progress_component.rb
@@ -1,5 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-class DepositProgressComponent < ApplicationComponent
-end
