--- app/components/work_form_component.html.erb
@@ -1,19 +1,0 @@
-<div id="editor">
-  <h1 class="row">Deposit your work</h1>
-
-  <%= form_with model: work, url: [work.collection, work], data: { controller: 'progress' },
-      html: { class: 'needs-validation row', novalidate: true, multipart: true } do |f| %>
-    <%= f.hidden_field :work_type %>
-    <%= f.hidden_field :subtype %> 
-    <%= render DepositProgressComponent.new %>
-    <%= render AddFilesComponent.new(form: f) %>
-    <%= render TitleComponent.new(form: f) %>
-    <%= render ContributorsComponent.new %>
-    <%= render DatesComponent.new(form: f) %>
-    <%= render DescriptionComponent.new(form: f) %>
-    <%= render EmbargoComponent.new(form: f) %>
-    <%= render LicenseComponent.new(form: f) %>
-    <%= render AgreementComponent.new(form: f) %>
-    <%= render ButtonsComponent.new(form: f) %>
-  <% end %>
-</div>
