--- app/components/work_form_component.rb
@@ -1,11 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-class WorkFormComponent < ApplicationComponent
-  attr_reader :work
-
-  sig { params(work: Work).void }
-  def initialize(work:)
-    @work = work
-  end
-end
