+++ app/components/works/contributors_component.html.erb
