+++ app/components/works/deposit_progress_component.html.erb
