+++ app/components/works/deposit_progress_component.rb
@@ -1,0 +1,7 @@
+# typed: strict
+# frozen_string_literal: true
+
+module Works
+  class DepositProgressComponent < ApplicationComponent
+  end
+end
