+++ app/components/works/description_component.html.erb
