+++ app/components/works/embargo_component.html.erb
