+++ app/components/works/embargo_component.rb
@@ -1,0 +1,12 @@
+# typed: true
+# frozen_string_literal: true
+
+module Works
+  class EmbargoComponent < ApplicationComponent
+    def initialize(form:)
+      @form = form
+    end
+
+    attr_reader :form
+  end
+end
