+++ app/components/works/license_component.html.erb
