+++ app/components/works/title_component.html.erb
