+++ app/views/works/new.html.erb
@@ -1,5 +1,5 @@
 <main class="mb-3 mb-md-5" id="content">
   <div class="container px-4 px-md-3">
-    <%= render WorkFormComponent.new(work: @work) %>
+    <%= render Works::WorkFormComponent.new(work: @work) %>
   </div>
 </main>
