+++ spec/components/works/title_component_spec.rb
@@ -3,7 +3,7 @@
 
 require 'rails_helper'
 
-RSpec.describe TitleComponent do
+RSpec.describe Works::TitleComponent do
   let(:form) do
     instance_double(ActionView::Helpers::FormBuilder,
                     label: nil,
